[manuf] move provisioning fw to `base` directory
The provisioning firmware in this repo is a baseline reference firmware
and should be named as such. This partially addresses #23426, and
enables testing provisioning extensions in pre-silicon.
