+++ sw/device/silicon_creator/manuf/base/flash_info_permissions.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_SKUS_EARLGREY_A0_SIVAL_BRINGUP_FLASH_INFO_PERMISSIONS_H_
-#define OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_SKUS_EARLGREY_A0_SIVAL_BRINGUP_FLASH_INFO_PERMISSIONS_H_
+#ifndef OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_BASE_FLASH_INFO_PERMISSIONS_H_
+#define OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_BASE_FLASH_INFO_PERMISSIONS_H_
 
 #include "sw/device/lib/dif/dif_flash_ctrl.h"
 
@@ -40,4 +40,4 @@ dif_flash_ctrl_region_properties_t kFlashInfoPage4Permissions = {
     .rd_en = kMultiBitBool4True,
     .scramble_en = kMultiBitBool4True};
 
-#endif  // OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_SKUS_EARLGREY_A0_SIVAL_BRINGUP_FLASH_INFO_PERMISSIONS_H_
+#endif  // OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_BASE_FLASH_INFO_PERMISSIONS_H_
