+++ sw/host/provisioning/cp/BUILD
@@ -11,7 +11,7 @@ rust_binary(
     testonly = True,
     srcs = ["src/main.rs"],
     data = [
-        "//sw/device/silicon_creator/manuf/skus/earlgrey_a0/sival_bringup:sram_cp_provision",
+        "//sw/device/silicon_creator/manuf/base:sram_cp_provision",
         "//third_party/openocd:jtag_cmsis_dap_adapter_cfg",
         "//third_party/openocd:jtag_olimex_cfg",
         "//third_party/openocd:openocd_bin",
